moving Main component into App container
